switch to useDispatch

+++ src/Components/Admin/Users/UserItem.jsx
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import { connect } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { useMutation } from '@apollo/react-hooks';
 import DELETE_USER from 'Mutations/deleteUser';
 import { showModal, WARNING_MODAL } from 'Redux/Actions/modalActions';
@@ -11,18 +11,18 @@ type Props = {
     username: string,
     id: number,
     admin: boolean,
-    sModal: Function
 };
 
-const UserItem = ({ sModal, username, id, admin }: Props) => {
+const UserItem = ({ username, id, admin }: Props) => {
     const [deleteUser, { data }] = useMutation(DELETE_USER);
+    const dispatch = useDispatch();
 
     const toggleModal = () => {
-        sModal(WARNING_MODAL, {
+        dispatch(showModal(WARNING_MODAL, {
             title: 'Delete user',
             message: `Are you sure you want to delete the user '${username}'?`,
             confirm: () => deleteUser({ variables: { id } }),
-        });
+        }));
     };
 
     return (
@@ -33,11 +33,4 @@ const UserItem = ({ sModal, username, id, admin }: Props) => {
     );
 };
 
-const mapDispatchToProps = (dispatch) => ({
-    sModal: (type, props) => dispatch(showModal(type, props)),
-});
-
-export default connect(
-    null,
-    mapDispatchToProps,
-)(UserItem);
+export default UserItem;
